Added missing `implements` and access modifiers in `BaseNavigationModeFactory`.  Rename the file to base-navigation-mode-factory.<...> to match the class name.
